fix: 修复v1.8.22编译错误 — 类型/模块引用修正
- stitching.rs: na::Matrix3 → Matrix3（已通过use导入）
- export_processing.rs: unwrap_or_else中的return → match+continue
- file_management.rs: String::new() → match+continue（正确的类型安全回退）
- lens_correction.rs: f64::MAX → f32::MAX（匹配focal字段类型）

+++ src-tauri/src/lens_correction.rs
@@ -324,7 +324,7 @@ impl Lens {
                 extract_dist_params(exact)
             } else if focal_length < distortions[0].focal {
                 extract_dist_params(distortions[0])
-            } else if focal_length > distortions.last().map(|d| d.focal).unwrap_or(f64::MAX) {
+            } else if focal_length > distortions.last().map(|d| d.focal).unwrap_or(f32::MAX) {
                 extract_dist_params(distortions.last().unwrap_or(&distortions[0]))
             } else {
                 let mut res = (0.0, 0.0, 0.0, 0);
